Add server  update and delete one item
