--- server/src/controllers/eCom.controller.js
@@ -1,31 +1,0 @@
-// create todo
-const stripe = require('stripe')('sk_test_LTQPhkCZjMU3La6mWPYzmJ6L00qXdcCpaO');
-const checkout = async (req, res) => {
-    console.log(req.body);
-    const items = req.body.items;
-    let lineItems = [];
-    items.forEach((item)=> {
-        lineItems.push(
-            {
-                price: item.id,
-                quantity: item.quantity
-            }
-        )
-    });
-
-    const session = await stripe.checkout.sessions.create({
-        line_items: lineItems,
-        mode: 'payment',
-        success_url: "http://localhost:3000/success",
-        cancel_url: "http://localhost:3000/cancel"
-    });
-
-    res.send(JSON.stringify({
-        url: session.url
-    }));
-};
-
-
-module.exports = {
-    checkout
-}
