--- server/src/routes/eCom.route.js
@@ -1,9 +1,0 @@
-const router = require("express").Router();
-
-const {
-    checkout
-} = require("../controllers/eCom.controller");
-
-router.post('/checkout', checkout);
-
-module.exports = router;
